move I/Message, eliminate Messaging.Tcp namespace, remove deprecated enums
