+++ src/Soulseek/Messaging/IMessageConnection.cs
@@ -10,7 +10,7 @@
 //     You should have received a copy of the GNU General Public License along with this program. If not, see https://www.gnu.org/licenses/.
 // </copyright>
 
-namespace Soulseek.Messaging.Tcp
+namespace Soulseek.Messaging
 {
     using System;
     using Soulseek.Tcp;
